requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
black==21.12b0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
cycler==0.10.0
discord.py==1.7.3
flake8==4.0.1
frozenlist==1.2.0
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.3
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.21.1
pandas==1.3.2
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
tomli==1.2.3
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
